// tmd/types/types.tsx
export type Font = {
  fontFamily: string;
  fontWeight?:
    | "normal"
    | "bold"
    | "100"
    | "200"
    | "300"
    | "400"
    | "500"
    | "600"
    | "700"
    | "800"
    | "900";
};

export type BaseFonts = {
  bold: Font;
  semiBold: Font;
  regular: Font;
  medium: Font;
  light: Font;
  thin: Font;
};

export type Fonts = {
  sans: BaseFonts;
  seasons: BaseFonts;
}